- Add Controller for creating method calls to cameras
The abstract Controller class is implemented by the CameraController class, which in turn has a method for sending commands to the cameras. The method checks if the command given in the argument is valid (i.e. in the ArrayList of valid commands) and tries to send the commands to the camera via the Connector class.

- Change name of Controller class to Connector (for obvious reasons)